Add files via upload
These are v. simple simulations to illustrate how investment returns of different assets with different means and standard deviations can compound over time. They're incredibly simple and unrealistic, but do provide some useful intuition. Two of the biggest limitations are a) it assumes independence of returns for equities and bonds and b) it assumes normally distributed returns.

It'd be pretty easy to adapt this to be more realistic. However, the key point here is that USS have not published the details of analysis even this sophisticated. When it comes to our science and out research, we rightly high levels of reproducibility, transparency and evidence. We should expect the same standards from the people who manage billions of pounds of assets on our behalf.